+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/azure
 
+## 4.0.28
+
+### Patch Changes
+
+- Updated dependencies [5fc7da5]
+- Updated dependencies [93b2acd]
+  - @ai-sdk/provider-utils@5.0.18
+  - @ai-sdk/deepseek@3.0.19
+  - @ai-sdk/openai@4.0.27
+
 ## 4.0.27
 
 ### Patch Changes