+++ packages/harness/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/harness
 
+## 1.0.54
+
+### Patch Changes
+
+- ai@7.0.48
+
+## 1.0.53
+
+### Patch Changes
+
+- bdde5d9: fix(harness): avoid placing harness bootstrap files in `/tmp` and instead use sandbox working directory
+- Updated dependencies [5fc7da5]
+- Updated dependencies [93b2acd]
+  - @ai-sdk/provider-utils@5.0.18
+  - ai@7.0.47
+
+## 1.0.52
+
+### Patch Changes
+
+- ai@7.0.46
+
 ## 1.0.51
 
 ### Patch Changes