+++ packages/otel/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/otel
 
+## 1.0.48
+
+### Patch Changes
+
+- ai@7.0.48
+
+## 1.0.47
+
+### Patch Changes
+
+- ai@7.0.47
+
+## 1.0.46
+
+### Patch Changes
+
+- ai@7.0.46
+
 ## 1.0.45
 
 ### Patch Changes