+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,24 @@
 
 ### Patch Changes
 
+- ai@7.0.48
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@7.0.47
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@7.0.46
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [d6ce0ee]
 - Updated dependencies [fa95504]
 - Updated dependencies [349afe7]